+++ src/main/java/com/github/siboxd/fatturapa/model/invoicebody/general/Art73Type.java
@@ -1,8 +1,11 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.invoicebody.general;
 
 import org.simpleframework.xml.Namespace;
 import org.simpleframework.xml.Root;
 
+/**
+ * It serves to direct control over the uniqueness of the numbering of documents
+ */
 @Root(name = "Art73Type")
 @Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
 public enum Art73Type {
